requirements.txt
appdirs==1.4.4
blowfish==0.6.1
distlib==0.3.1
filelock==3.0.12
kafka-python==2.0.2
llvmlite==0.36.0
numba==0.53.1
numpy==1.20.2
Pillow==8.2.0
six==1.15.0
urllib3==1.26.4
virtualenv==20.4.4
